fix: build script
